+++ radicale/hook/email/__init__.py
@@ -1,6 +1,7 @@
 # This file is related to Radicale - CalDAV and CardDAV server
 # for email notifications
 # Copyright © 2025-2025 Nate Harris
+# Copyright © 2026-2026 Peter Bieringer <pb@bieringer.de>
 #
 # This library is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -58,6 +59,10 @@ PLUGIN_CONFIG_SCHEMA = {
             "type": str
         },
         "smtp_password": {
+            "value": "",
+            "type": str
+        },
+        "smtp_password_file": {
             "value": "",
             "type": str
         },
@@ -914,13 +919,21 @@ def _read_event(vobject_data: str) -> EmailEvent:
 class Hook(BaseHook):
     def __init__(self, configuration):
         super().__init__(configuration)
+        smtp_password = self.configuration.get("hook", "smtp_password")
+        smtp_password_file = self.configuration.get("hook", "smtp_password_file")
+        if smtp_password_file:
+            try:
+                with open(smtp_password_file, "r", encoding="utf-8") as file:
+                    smtp_password = file.read().rstrip("\n")
+            except Exception as e:
+                logger.error("Email hook option 'smtp_password_file' provided but not readable: %r (%r)", smtp_password_file, e)
         self.email_config = EmailConfig(
             host=self.configuration.get("hook", "smtp_server"),
             port=self.configuration.get("hook", "smtp_port"),
             security=self.configuration.get("hook", "smtp_security"),
             ssl_verify_mode=self.configuration.get("hook", "smtp_ssl_verify_mode"),
             username=self.configuration.get("hook", "smtp_username"),
-            password=self.configuration.get("hook", "smtp_password"),
+            password=smtp_password,
             from_email=self.configuration.get("hook", "from_email"),
             send_mass_emails=self.configuration.get("hook", "mass_email"),
             dryrun=self.configuration.get("hook", "dryrun"),